arteria_samplesheet.csv -> Samplesheet.csv
We need the samplesheet to specifically be named Samplesheet.csv
downstream - and this was the easiest way to fix this. For extra safety
the old samplesheet will now be copied to a file tagged with the date and
time so that any samplesheet written there by e.g. the sequencer isn't
lost.